% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micnorm.R
\name{add_activity_string}
\alias{add_activity_string}
\title{Add string describing students' activity.}
\usage{
add_activity_string(students)
}
\arguments{
\item{students}{(tibble) students}
}
\value{
a tibble with same columns as students + activity_string
}
\description{
\code{add_activity_string()} adds a string describing students' activity to
students' tibble.
}
