% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micnorm.R
\name{augment_points}
\alias{augment_points}
\title{Augment activity points for different number of call ups and for illness.}
\usage{
augment_points(students)
}
\arguments{
\item{students}{(tibble) students}
}
\value{
a tibble with same columns as students + raw_points
}
\description{
\code{augment_points()} augments activity points for different number of call ups.
}
