% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micnorm.R
\name{format_number}
\alias{format_number}
\title{Format number.}
\usage{
format_number(x)
}
\arguments{
\item{x}{(number) number to be formatted}
}
\value{
a number
}
\description{
\code{format_number()} formats a number for output string---it replaces NA and
Inf with 0 and rounds the number to 3 decimal places.
}
