% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micnorm.R
\name{get_alternative_attendance}
\alias{get_alternative_attendance}
\title{Get alternative attendance points from IS.}
\usage{
get_alternative_attendance(creds, alt_attendance_notebook)
}
\arguments{
\item{alt_attendance_notebook}{(string) name of the notebook where
alternative attendance points are stored}

\item{...}{credentials}
}
\value{
a tibble with following columns:
\itemize{
\item course,
\item student_uco,
\item alt_attendance_points
}
}
\description{
\code{get_alternative_attendance()} downloads alternative attendance points
from IS.
}
