% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micnorm.R
\name{get_attendance}
\alias{get_attendance}
\title{Get and normalize attendance in several courses.}
\usage{
get_attendance(
  creds,
  no_of_seminars,
  max_points_attendance,
  alt_attendance_notebook
)
}
\arguments{
\item{...}{credentials separated by commas}
}
\value{
a tibble with following columns:
\itemize{
\item course,
\item student_uco,
\item attendance_string,
\item attendance_points,
\item normalized_attendance, and
\item credentials
}
}
\description{
Get and normalize attendance in several courses.
}
