% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micnorm.R
\name{get_list_of_existing_notebooks}
\alias{get_list_of_existing_notebooks}
\title{Get list of all activity points notebooks.}
\usage{
get_list_of_existing_notebooks(creds, name_mask = "^bodysemin\\\\d{2}$")
}
\arguments{
\item{name_mask}{(string) regular expression that says how are named
the notebooks that include activity points}

\item{`...`}{credentials; separated by commas}
}
\value{
a tibble of names of all existing blocks; it includes columns:
\itemize{
\item course,
\item name,
\item shortcut, and
\item credential
}
}
\description{
\code{get_list_of_existing_notebooks()} downloads names of all existing blocks
within several courses
}
\details{
it is protected against errors; it logs
}
\examples{
\dontrun{
blocks <- get_list_of_existing_notebooks(micprez, mivs)
}
}
