% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micnorm.R
\name{get_names_of_all_existings_blocks}
\alias{get_names_of_all_existings_blocks}
\title{Get list of all seminar points notebooks.}
\usage{
get_names_of_all_existings_blocks(...)
}
\arguments{
\item{`...`}{credentials; separated by commas}
}
\value{
a tibble of names of all existing blocks; it includes columns:
\itemize{
\item course,
\item name,
\item shortcut, and
\item credential
}
}
\description{
\code{get_names_of_all_existings_blocks()} downloads names of all existing blocks
within several courses
}
\details{
it is protected against errors; it logs
}
\examples{
\dontrun{
blocks <- get_names_of_all_existings_blocks(micprez, mivs)
}
}
