% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micnorm.R
\name{get_renegades}
\alias{get_renegades}
\title{Get list of students that stopped studying.}
\usage{
get_renegades(creds)
}
\arguments{
\item{`...`}{credentials; separated by commas}
}
\value{
an integer vectors including UCOs of all renegades
}
\description{
\code{get_renegades()} downloads names of all renegades (i.e. students that
stopped working within several courses.
}
\details{
it is protected against errors; it logs
}
\examples{
\dontrun{
renegades <- get_renegades(micprez, mivs)
}
}
