% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micnorm.R
\name{get_students_attached_to_teachers}
\alias{get_students_attached_to_teachers}
\title{Get all students and attach them to teachers.}
\usage{
get_students_attached_to_teachers(creds)
}
\arguments{
\item{...}{credentials}
}
\value{
a tibble with the following columns:
\itemize{
\item course, seminar,
\item teacher_uco, teacher_last_name, teacher_first_name,
\item student_uco, student_last_name, student_first_name,
\item credentials
}
}
\description{
Get all students and attach them to teachers.
}
