% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{is_same}
\alias{is_same}
\title{Check if two queries to IS gives the same normalized data.}
\usage{
is_same(x, y)
}
\arguments{
\item{x}{tibble returned by \code{\link{normalize_micro}}}

\item{y}{tibble returned by \code{\link{normalize_micro}}}
}
\value{
A logical value indicating whether the two tibbles are the same
}
\description{
This function checks if two queries to IS give the same data.
As IS returns the data in different order, the data frames are
sorted by \code{student_uco} before comparison.
}
\examples{
res <- normalize_micro(
  micro,
  mivs,
  course_mapping = mivsmicro,
  export_to_IS = FALSE,
  send_mail = FALSE
)
res2 <- normalize_micro(
  micro,
  mivs,
  course_mapping = mivsmicro,
  export_to_IS = FALSE,
  send_mail = FALSE
)
is_same(df1, df2) # should return TRUE
}
