% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micnorm.R
\name{normalize_points}
\alias{normalize_points}
\title{Normalize points for activity on seminar.}
\usage{
normalize_points(
  students,
  max_points_activity,
  activity_const_a,
  activity_const_b
)
}
\arguments{
\item{students}{(tibble)}

\item{max_points_activity}{(number) maximum normalized points
for activity}

\item{activity_const_a}{(number) normalization constant A}

\item{activity_const_b}{(number) normalization constant B}
}
\value{
a tibble
}
\description{
Normalize points for activity on seminar.
}
