% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micnorm.R
\name{read_one_test}
\alias{read_one_test}
\title{Read test points from a notebook.}
\usage{
read_one_test(uco_cred, type, test, year, course = "BPE_MIE1")
}
\arguments{
\item{uco_cred}{(list) credentials of the user created with
\code{MUIS::uco_credentials()}}

\item{type}{(string) type of the test; one of "abcd" or "tft"}

\item{test}{(string) test number including the letter; e.g. "2B"}

\item{year}{(string) year of the course; e.g. "2024"}

\item{course}{(string) name of the base course; the default is "BPE_MIE1"}
}
\value{
a tibble with columns:
\itemize{
\item uco,
\item student_name
\item points
\item time
}
}
\description{
\code{read_test_points()} reads test points from a notebook.
}
\examples{
\dontrun{
read_test_points(uco_cred, "abcd", "2B", "2024")
}
}
