% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micnorm.R
\name{read_points_from_blocks}
\alias{read_points_from_blocks}
\title{Read studnets' points from seminar notebooks.
\code{read_points_from_blocks()} reads all blocks found by
get_names_of_all_existings_blocks() and returns students' points stored there}
\usage{
read_points_from_blocks(blocks)
}
\arguments{
\item{blocks}{a tibble returned by \code{get_names_of_all_existings_blocks()}}
}
\value{
a tibble with columns:
\itemize{
\item uco,
\item points,
\item course, and
\item block
}
}
\description{
Read studnets' points from seminar notebooks.
\code{read_points_from_blocks()} reads all blocks found by
get_names_of_all_existings_blocks() and returns students' points stored there
}
