% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micnorm.R
\name{safely_create_normalized_block}
\alias{safely_create_normalized_block}
\title{Create normalization notebooks.}
\usage{
safely_create_normalized_block(name, shortcut, ...)
}
\arguments{
\item{name}{(string) full name of the notebook}

\item{shortcut}{(string) shortcut name of the notebook}

\item{...}{credentials}
}
\value{
none, it only creates the blocks in IS
}
\description{
\code{safely_create_normalized_block()} creates notebooks for normalized points
for several courses if they don't exist.
}
