% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micnorm.R
\name{send_mail}
\alias{send_mail}
\title{Send email}
\usage{
send_mail(subject, body, sender = "847@muni.cz", recipients = sender)
}
\arguments{
\item{subject}{(character string) mail header}

\item{body}{(character string nebo MIME) mail body (can include \n)}

\item{sender}{(character string) sender's email address; implicitly mine}

\item{recipient}{(character string) recipient's email address;
implicitly the same as sender's one}
}
\value{
none -- it just sends a mail
}
\description{
\code{send_mail()} sends an mail from my school mail account to my school mail
account
}
\details{
there must be sendmail present on the server
}
