% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micnorm.R
\name{start_logging}
\alias{start_logging}
\title{Start logging.}
\usage{
start_logging(log_folder, norm_block)
}
\arguments{
\item{log_folder}{(string) path to folder where logs are written}

\item{norm_block}{(string) shortcut of notebook with normalized points}
}
\value{
path to log_file
}
\description{
\code{start_logging()} creates a log folder if necessary and starts logging into
a file. It also logs the start of the script.
}
