% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micnorm.R
\name{test_url}
\alias{test_url}
\title{Return URL of a test points notebook.}
\usage{
test_url(type, test, year, course)
}
\arguments{
\item{type}{(string) type of the test; one of "abcd" or "tft"}

\item{test}{(string) test number including the letter; e.g. "2B"}

\item{year}{(string) year of the course; e.g. "2024"}

\item{course}{(string) name of the base course; e.g. "BPE_MIE1"}
}
\value{
(string) URL of the test points notebook
}
\description{
\code{test_url()} returns URL of a test points notebook.
}
\examples{
\dontrun{
test_url("abcd", "2B", "2024", "BPE_MIE1")
}
}
