% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micnorm.R
\name{unite_courses}
\alias{unite_courses}
\title{If there are more courses, join their seminars.}
\usage{
unite_courses(students, mapping)
}
\arguments{
\item{students}{(tibble) students}

\item{mapping}{(tibble) mapping of courses; must include columns:
\itemize{
\item course,
\item source_seminar, and
\item target_seminar
}}
}
\value{
a tibble with the same columns as \code{students}
}
\description{
\code{unite_courses()} joins seminars of several courses if they are the same
(e.g. BPE_MIE1 and MPE_MIVS)
}
\details{
if \code{mapping} is \code{NULL}, it returns \code{students} unchanged
}
