% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{view_output}
\alias{view_output}
\title{View Output}
\usage{
view_output(tab, semin = NULL, by_semin = TRUE)
}
\arguments{
\item{tab}{tibble provided by \code{\link{normalize_micro}}.}

\item{semin}{vector of seminars to be viewed provided as strings,
e.g., \code{"04"}.}

\item{by_semin}{logical, if \code{TRUE}, the output is sorted by
seminar first, and then by last names; \code{TRUE} is default
value.}
}
\value{
None
}
\description{
This function is used to view the output string of Microeconomics 1
normalization.
}
\examples{
view_output(tab = my_table)
view_output(tab = my_table, semin = c("04", "06"))
}
