% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micnorm.R
\name{write_data_to_is}
\alias{write_data_to_is}
\title{Writes normalized points to IS.}
\usage{
write_data_to_is(students, norm_block, ...)
}
\arguments{
\item{students}{... a tibble with points}

\item{norm_block}{... shortcut name of the block}

\item{...}{credentials}
}
\value{
none, it only writes the data to the blocks in IS
}
\description{
\code{write_data_to_is()} writes students' normalized points to normalization
blocks in IS for several courses.
}
