% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micnorm.R
\name{add_attendance}
\alias{add_attendance}
\title{Add attendance data to a table}
\usage{
add_attendance(
  tab,
  ...,
  no_of_seminars,
  max_points_attendance,
  alt_attendance_notebook
)
}
\arguments{
\item{tab}{A tibble containing the input table.}

\item{...}{credentials separated by commas.}

\item{no_of_seminars}{The total number of seminars.}

\item{max_points_attendance}{The maximum points that can be obtained
for attendance.}

\item{alt_attendance_notebook}{The name of the notebook with an alternative
attendance.}
}
\value{
A tibble containing the input table with attendance data added.
}
\description{
This function adds attendance data to a table.
}
