% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micnorm.R
\name{add_output_string}
\alias{add_output_string}
\title{Add string describing students' activity and attendance.}
\usage{
add_output_string(students, max_points_attendance, max_points_activity)
}
\arguments{
\item{students}{(tibble) students}
}
\value{
a tibble with same columns as students + output_string
}
\description{
\code{add_output_string()} adds a string describing students' activity
and attendance to students' tibble.
}
