% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micnorm.R
\name{create_and_send_mail}
\alias{create_and_send_mail}
\title{Create email from log and send it.}
\usage{
create_and_send_mail()
}
\value{
none -- it only sends the mail
}
\description{
Create email from log and send it.
}
