% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micnorm.R
\name{get_activity_points}
\alias{get_activity_points}
\title{Get and process students' activity points.}
\usage{
get_activity_points(..., name_mask = "^bodysemin\\\\d{2}$")
}
\arguments{
\item{...}{credentials}

\item{name_mask}{(string) regular expression that says how are named
the notebooks where activity points are stored}
}
\value{
some tibble
}
\description{
Get and process students' activity points.
}
