% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micnorm.R
\name{get_activity_points}
\alias{get_activity_points}
\title{Get and process students' activity points.}
\usage{
get_activity_points(...)
}
\arguments{
\item{...}{credentials}
}
\value{
some tibble
}
\description{
Get and process students' activity points.
}
