% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micnorm.R
\name{get_all_students}
\alias{get_all_students}
\title{Get list of all students.}
\usage{
get_all_students(...)
}
\arguments{
\item{`...`}{credentials; separated by commas}
}
\value{
a tibble of all students; it includes columns:
\itemize{
\item course,
\item seminar,
\item student_uco,
\item student_last_name,
\item student_first_name, and
\item credentials
}
}
\description{
\code{get_all_students()} downloads a list of all students within all seminars of
several courses
}
\details{
it is protected against errors; it logs
}
\examples{
\dontrun{
students <- get_all_students(micprez, mivs)
}
}
