% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micnorm.R
\name{get_all_teachers}
\alias{get_all_teachers}
\title{Get list of all teachers.}
\usage{
get_all_teachers(...)
}
\arguments{
\item{`...`}{credentials; separated by commas}
}
\value{
a tibble of all teachers; it includes columns:
\itemize{
\item course,
\item seminar,
\item teacher_uco,
\item teacher_last_name, and
\item teacher_first_name
}
}
\description{
\code{get_all_teachers()} downloads a list of all teachers within several courses
}
\details{
it is protected against errors; it logs
}
\examples{
\dontrun{
teachers <- get_all_teachers(micprez, mivs)
}
}
