% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micnorm.R
\name{get_attendance}
\alias{get_attendance}
\title{Get attendance in several courses.}
\usage{
get_attendance(...)
}
\arguments{
\item{...}{credentials separated by commas}
}
\description{
Get attendance in several courses.
}
