% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micnorm.R
\name{normalize_micro}
\alias{normalize_micro}
\title{Read IS notebooks, normalize students' points, and write it back to IS.}
\usage{
normalize_micro(
  ...,
  course_mapping = NULL,
  no_of_seminars = 12L,
  max_points_attendance = 6,
  max_points_activity = 24,
  activity_const_a = 20,
  activity_const_b = 120,
  activity_name_mask = "^bodysemin\\\\d{2}$",
  alt_attendance_notebook = "preznahr",
  norm_name = "Normované body za průběžnou práci na semináři",
  norm_block = "bodsemin",
  log_folder = "logs",
  group_file_name = "last_groupings.RData",
  sender = "847@muni.cz",
  recipient = sender,
  export_to_IS = TRUE,
  send_mail = TRUE
)
}
\arguments{
\item{course_mapping}{(optional, tibble) mapping of courses; must include
columns: course, source_seminar, and target_seminar}

\item{no_of_seminars}{(optional, integer) number of seminars in the given
year; implicit value is 12L}

\item{max_points_attendance}{(optional, number) maximum normalized points
for attendance; implicit value is 6}

\item{max_points_activity}{(optional, number) maximum normalized points
for activity; implicit value is 24}

\item{activity_const_a}{(optional, number) normalization constant A;
implicit value is 20}

\item{activity_const_b}{(optional, number) normalization constant B;
implicit value is 120}

\item{activity_name_mask}{(optional, string) regular expression that says
how are named the notebooks where activity points are stored;
implicitly "^bodysemin\\d{2}$",}

\item{norm_name}{(optional, string) name of notebook with normalized points}

\item{norm_block}{(optional, string) shortcut of notebook with normalized
points}

\item{log_folder}{(optional, string) path to folder where logs are written;
if it doesn't exist, it is created}

\item{group_file_name}{(optional, string) name of file, where partial data
are written}

\item{sender}{(optional, string) email address of sender}

\item{recipient}{(optional, string) email address of recipient}

\item{export_to_IS}{(optional, logical) whether to export the data to IS;
implicit value is TRUE}

\item{`...`}{credentials separated by commas}
}
\value{
(silently) tibble with data created
}
\description{
Read IS notebooks, normalize students' points, and write it back to IS.
}
