% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micnorm.R
\name{normalize_micro}
\alias{normalize_micro}
\title{Read IS notebooks, normalize students' points, and write it back to IS.}
\usage{
normalize_micro(
  ...,
  no_of_seminars = 12L,
  max_points_attendance = 6,
  max_points_activity = 24,
  activity_const_a = 20,
  activity_const_b = 120,
  norm_name = "Normované body za průběžnou práci na semináři",
  norm_block = "bodsemin",
  log_folder = "logs",
  group_file_name = "last_groupings.RData"
)
}
\arguments{
\item{no_of_seminars}{(optional, integer) number of seminars in the given
year; implicit value is 12L}

\item{max_points_attendance}{(optional, number) maximum normalized points
for attendance; implicit value is 6}

\item{max_points_activity}{(optional, number) maximum normalized points
for activity; implicit value is 24}

\item{activity_const_a}{(optional, number) normalization constant A;
implicit value is 20}

\item{activity_const_b}{(optional, number) normalization constant B;
implicit value is 120}

\item{norm_name}{(optional, string) name of notebook with normalized points}

\item{norm_block}{(optional, string) shortcut of notebook with normalized
points}

\item{log_folder}{(optional, string) path to folder where logs are written;
if it doesn't exist, it is created}

\item{group_file_name}{(optional, string) name of file, where partial data
are written}

\item{`...`}{credentials separated by commas}
}
\value{
(silently) tibble with data created
}
\description{
Read IS notebooks, normalize students' points, and write it back to IS.
}
